requirements.txt
livekit-agents[openai,silero,deepgram,cartesia]~=1.0
python-dotenv>=1.0.0
